Move set to sys crate
